requirements.txt
flask==3.0.3
flask-cors==4.0.1
pandas==2.2.3
gunicorn==22.0.0
numpy==1.26.4
google-generativeai==0.7.2